remove mountpoint as soon as possible
Currently the mountpoint for a volume was only deleted when the volume
itself got deleted. Now we remove the mountpoint if there is no active
mount for that volume. This avoids stale mount directories in case where
for example a volume is created on one host and then removed on another.