allow removal of volumes if they have not been used
Currently a volume could only be removed if it was actively used by this
host (i.e., self.mounts[name] exists). Relax that and allow removal if
the volume exists and is not somewhere in Primary role. This for example
helps on simple reboots, where so far it was necessary to use the volume
once before removal was even possible.