rpm: Work around rpm stupidity
On RHEL people might install docker from the RHEL repos or from the
docker web site. RHEL names docker "docker", official docker(.com) names
the package "docker-engine".

As usual, rpm sucks, it is not possible to specify an OR dependency like
it is possible for Debian packages.

Hack around that stupidity in the "%pre" section and fail if none of the
docker packages is installed. A proper fix would be to provide a
virtual packages which provides "TheDockerService", which is provides by
RHEL and official docker.